fix: noteのfeedが取れないバグ修正
